Initial version of my Tetris-game.
I consider it a MVP for this project, being able to play, but not having much elements to it. Code is still very much WIP and only slight refactors applied. Comment pass done with some refactor ideas and describing the logic. I did watch some YouTube guides to save time on the very basics of Tetris logic because it has already been done so many times. However I did not watch any guides that were done with C#, I only watched GDScript / logic related guides. All of the assets are self made with Gimp as well.